requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
arabic-reshaper==2.1.3
asn1crypto==0.24.0
astroid==1.5.3
autopep8==1.3
Babel==2.7.0
backports-abc==0.5
backports.functools-lru-cache==1.4
bcrypt==3.1.4
billiard==3.3.0.23
boto3==1.4.4
botocore==1.5.26
celery==3.1.25
celery-progress==0.0.9
cffi==1.11.2
colorama==0.3.9
configparser==3.5.0
coreapi==2.1.1
cryptography==2.1.4
diff-match-patch==20121119
Django==1.9.11
django-admin-list-filter-dropdown==1.0.1
django-admin-rangefilter==0.4.0
django-admin-view-permission==1.6
django-braces==1.4.0
django-celery==3.2.1
django-celery-results==1.0.1
django-cors-headers==2.0.1
django-debug-toolbar==1.6
django-easy-pdf==0.1.1
django-import-export==0.5.1
django-jet==1.0.8
django-model-utils==3.0.0
django-oauth-toolkit==0.12.0
django-rest-swagger==2.1.1
django-s3direct==0.4.8
django-tinymce==2.6.0
djangorestframework==3.5.3
docutils==0.13.1
enum34==1.1.6
et-xmlfile==1.0.1
flower==0.9.3
future==0.18.2
futures==3.0.5
gunicorn==19.7.1
html5lib==1.1
idna==2.6
ipaddress==1.0.19
isort==4.2.15
itypes==1.1.0
jdcal==1.3
jmespath==0.9.2
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
meld3==1.0.2
oauthlib==2.0.1
odfpy==1.3.6
olefile==0.44
openapi-codec==1.2.1
openpyxl==2.4.7
packaging==16.8
paramiko==2.4.0
pep8==1.7.0
Pillow==4.0.0
psycopg2==2.6.2
psycopg2-binary==2.7.7
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pylint==1.7.2
PyNaCl==1.2.1
pyparsing==2.1.10
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
redis==2.10.5
reportlab==3.5.59
requests==2.13.0
s3transfer==0.1.10
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.2
supervisor==3.3.1
tablib==0.12.1
tornado==5.1.1
unicodecsv==0.14.1
uritemplate==3.0.0
vine==1.1.3
webencodings==0.5
wrapt==1.10.10
xhtml2pdf==0.2.5
xlrd==1.1.0
xlwt==1.2.0
django-environ==0.4.5